+++ src/algorithms.rs
@@ -4,6 +4,7 @@ mod mgf;
 
 pub(crate) mod generate;
 pub(crate) mod oaep;
+pub(crate) mod pad;
 pub(crate) mod pkcs1v15;
 pub(crate) mod pss;
 pub(crate) mod rsa;
